# packages/openapi.json
{
  "components": {
    "schemas": {
      "ActivityResponse": {
        "properties": {
          "count": {
            "type": "integer"
          },
          "events": {
            "items": {
              "additionalProperties": {},
              "type": "object"
            },
            "type": "array"
          }
        },
        "required": [
          "events",
          "count"
        ],
        "type": "object"
      },
...
      "OwnSessionsResponse": {
        "properties": {
          "count": {
            "type": "integer"
          },
          "sessions": {
            "items": {
              "additionalProperties": {},
              "type": "object"
            },
            "type": "array"
          }
        },
        "required": [
          "sessions",
          "count"
        ],
        "type": "object"
      },
...
      "ProfileResponse": {
        "properties": {
          "field_permissions": {
            "additionalProperties": {
              "additionalProperties": {},
              "type": "object"
            },
            "type": "object"
          },
          "identity": {
            "additionalProperties": {},
            "type": "object"
          },
          "schema": {
            "additionalProperties": {},
            "type": "object"
          }
        },
        "required": [
          "identity",
          "schema",
          "field_permissions"
        ],
        "type": "object"
      },
...
      "StatusResponse": {
        "properties": {
          "status": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "type": "object"
      },
      "UpdateProfileRequest": {
        "properties": {
          "display_name": {
            "type": "string"
          },
          "profile": {
            "additionalProperties": {},
            "type": "object"
          }
        },
        "type": "object"
      },
      "UpdateProfileResponse": {
        "properties": {
          "fields_changed": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "fields_changed"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "PAT or session token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Identity and access management API. Schema-driven, extensible, open source.",
    "title": "Zitadel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/account/activity": {
      "get": {
        "operationId": "listOwnActivity",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityResponse"
                }
              }
            },
            "description": "List own activity"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List own activity",
        "tags": [
          "Account"
        ]
      }
    },
    "/v1/account/profile": {
      "get": {
        "operationId": "getProfile",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            },
            "description": "Get own profile"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get own profile",
        "tags": [
          "Account"
        ]
      },
      "patch": {
        "operationId": "updateProfile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateProfileResponse"
                }
              }
            },
            "description": "Update own profile"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update own profile",
        "tags": [
          "Account"
        ]
      }
    },
    "/v1/account/sessions": {
      "get": {
        "operationId": "listOwnSessions",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnSessionsResponse"
                }
              }
            },
            "description": "List own sessions"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List own sessions",
        "tags": [
          "Account"
        ]
      }
    },
    "/v1/account/sessions/revoke-others": {
      "post": {
        "operationId": "revokeOtherSessions",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "Revoke all other sessions"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Revoke all other sessions",
        "tags": [
          "Account"
        ]
      }
    },
    "/v1/account/sessions/{id}/revoke": {
      "post": {
        "operationId": "revokeOwnSession",
        "parameters": [
          {
            "description": "Resource ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "Revoke own session"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Revoke own session",
        "tags": [
          "Account"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "This instance",
      "url": "/"
    }
  ],
...
}